ZA use sayit functionality from master #1027
Point section-view urls to <full_slug> type ones.  This leads to urls
like

    /hansard/hansard/2013/05/08/proceedings-of-the-national-assembly-wednesday-08/appropriation-bill

where the repetition of /hansard is a little jarring.  I've not been
able to get this to work mounted at the top-level however.

The Committee-specific redirect now inherits from SectionView in order
to take advantage of its .get_object() parsing.